# apps/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /files:
    post:
      operationId: Files.upload
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponseDto'
      tags:
        - files
  /files/{id}:
    get:
      operationId: Files.read
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponseDto'
      tags:
        - files
    delete:
      operationId: Files.remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponseDto'
      tags:
        - files
...
info:
  title: Platform Core API
  description: Platform Core API documentation
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      in: header
  schemas:
...
    FileViewDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: notes.txt
        type:
          type: string
          example: text/plain
        size:
          type: number
          example: 10
        path:
          type: string
          example: memory/notes.txt
        url:
          type: string
          example: https://files.test/notes.txt
        uploadedById:
          type: string
      required:
        - id
        - name
        - type
        - size
        - path
        - url
        - uploadedById
    FileResponseDto:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: File uploaded
        data:
          $ref: '#/components/schemas/FileViewDto'
      required:
        - status
        - message
        - data
...